fix: UI情報はlibディレクトリで扱う用に修正
