+++ lib/constant/weather_condition.dart
@@ -1,6 +1,4 @@
-import 'package:flutter_svg/svg.dart';
-import 'package:flutter_training/gen/assets.gen.dart';
-
+/// 天気の種類を扱うEnum
 enum WeatherCondition {
   cloudy,
   sunny,
@@ -10,17 +8,4 @@ enum WeatherCondition {
       WeatherCondition.values.singleWhere(
         (element) => element.name == conditionName,
       );
-}
-
-extension WeatherTypePath on WeatherCondition {
-  SvgPicture get path {
-    switch (this) {
-      case WeatherCondition.cloudy:
-        return Assets.images.cloudy.svg();
-      case WeatherCondition.sunny:
-        return Assets.images.sunny.svg();
-      case WeatherCondition.rainy:
-        return Assets.images.rainy.svg();
-    }
-  }
 }
